Added texture class. Constructs a texture from the given file and flags, and can be used by its .use() member function.
Added TexturedCube derived class (cube). Essentially is a cube with an additional Texture member. And an overriden draw command.

Updated Cube class to be more inheritance-friendly.

Added freeFileData member function to FileReader.h

The program will now render a green cube, a blue cube, and a textured cube. (a.bmp must still be placed in the debug folder for this to work.)